chore(src): reorganize File elements
* hx-drop-fence
* hx-drop-zone
* hx-file-control
* hx-file-icon
* hx-file-input
* hx-file-tile
