+++ src/helix-ui/elements/hx-drop-fence/_shadow.less
@@ -1,4 +1,4 @@
-@import "./HXElement";
+@import "../HXElement";
 
 #hxDropFence {
   align-items: center;
