+++ src/helix-ui/elements/hx-drop-fence/index.js
@@ -1,6 +1,7 @@
-import { HXElement } from './HXElement';
-import shadowMarkup from './HXDropFenceElement.html';
-import shadowStyles from './HXDropFenceElement.less';
+import { HXElement } from '../HXElement.js';
+
+import shadowMarkup from './_shadow.html';
+import shadowStyles from './_shadow.less';
 
 /**
  * Defines behavior for the `<hx-drop-fence>` element.
