+++ src/helix-ui/elements/hx-file-icon/_shadow.html
