+++ src/helix-ui/elements/hx-file-icon/_shadow.less
@@ -1,4 +1,4 @@
-@import './HXElement';
+@import '../HXElement';
 
 #hxFileIcon {
   position: relative;
