+++ src/helix-ui/elements/hx-file-input/_shadow.less
@@ -1,4 +1,4 @@
-@import './HXElement';
+@import '../HXElement';
 
 #hxFileInput {
   //width: 100%;
