+++ src/helix-ui/elements/hx-file-input/index.js
@@ -1,6 +1,7 @@
-import { HXElement } from './HXElement';
-import shadowMarkup from './HXFileInputElement.html';
-import shadowStyles from './HXFileInputElement.less';
+import { HXElement } from '../HXElement.js';
+
+import shadowMarkup from './_shadow.html';
+import shadowStyles from './_shadow.less';
 
 /**
  * Defines behavior for the `<hx-file-input>` element.
