+++ src/helix-ui/elements/hx-file-tile/_shadow.less
@@ -1,4 +1,4 @@
-@import "./HXElement";
+@import "../HXElement";
 @import "base/links";
 @import "components/loaders";
 @import (reference) "components/buttons";
