+++ src/helix-ui/elements/index.js
@@ -8,14 +8,14 @@ export { HXCheckboxSetElement } from './hx-checkbox-set/index.js';
 export { HXDisclosureElement } from './HXDisclosureElement';
 export { HXDivElement } from './HXDivElement';
 export { HXDrawerElement } from './HXDrawerElement';
-export { HXDropFenceElement } from './HXDropFenceElement';
-export { HXDropZoneElement } from './HXDropZoneElement';
+export { HXDropFenceElement } from './hx-drop-fence/index.js';
+export { HXDropZoneElement } from './hx-drop-zone/index.js';
 export { HXElement } from './HXElement';
 export { HXErrorElement } from './hx-error/index.js';
-export { HXFileControlElement } from './HXFileControlElement';
-export { HXFileIconElement } from './HXFileIconElement';
-export { HXFileInputElement } from './HXFileInputElement';
-export { HXFileTileElement } from './HXFileTileElement';
+export { HXFileControlElement } from './hx-file-control/index.js';
+export { HXFileIconElement } from './hx-file-icon/index.js';
+export { HXFileInputElement } from './hx-file-input/index.js';
+export { HXFileTileElement } from './hx-file-tile/index.js';
 export { HXFormControlElement } from './HXFormControlElement';
 export { HXIconElement } from './HXIconElement';
 export { HXMenuElement } from './hx-menu/index.js';
